Atomic Designに則ったディレクトリをCopmonentsに作成しファイル移動
