+++ src/App.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import HeaderComponent from './components/HeaderComponent.vue'
+import HeaderComponent from './components/organisms/HeaderComponent.vue'
 </script>
 <template>
     <HeaderComponent />
